added typescript support and fixed build dist
