+++ frontend/.gitignore
@@ -9,7 +9,8 @@ lerna-debug.log*
 
 node_modules
 dist
-dist-ssr
+dist-react
+dist-electron
 *.local
 
 # Editor directories and files
