--- frontend/src/components/MainContent.jsx
@@ -1,21 +1,0 @@
-import { useState } from 'react'
-import '../stylesheets/Navbar.css'
-
-class MainContent extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-    }
-  }
-
-  render() {
-    <div class="sidebar">
-        <ul>
-            <li>Dashboard</li>
-            <li>Market</li>
-            <li>Upload</li>
-            <li>Wallet/Mining</li>
-        </ul>
-    </div>
-  }
-}
