+++ frontend/tsconfig.app.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./src/ui/app.tsx","./src/ui/main.tsx","./src/ui/vite-env.d.ts"],"version":"5.6.2"}
