+++ frontend/vite.config.ts
@@ -3,6 +3,9 @@ import react from "@vitejs/plugin-react";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  server: { port: 3000 },
   plugins: [react()],
+  base: "./",
+  build: {
+    outDir: "dist-react",
+  },
 });
